requirements.txt
fastapi
uvicorn
jinja2
requests
redis
sqlalchemy==2.0.36
python-dotenv
mysql-connector-python==8.4.0
mariadb
python-multipart
pyjwt
python-jose
passlib
matplotlib
plotly
pyotp
nano
pandas